+++ SlotMachienNXT/src/slotmachien/handlers/SMMotorHandler.java
@@ -6,8 +6,8 @@ import java.util.Queue;
 import lejos.nxt.NXTRegulatedMotor;
 import observable.AbstractObservable;
 import slotmachien.Command;
-import slotmachien.MotorBlock;
 import slotmachien.Position;
+import slotmachien.internal.MotorBlock;
 import slotmachien.signals.MovedToSignal;
 
 public class SMMotorHandler extends AbstractObservable<MovedToSignal> {
